DBA Names and Facility Cleanup
The code is the same in the two files. The python file will replace your food inspections file with the new one. Also added some facility names to some places that did not have them. Facilities have NAN if it does not have a value. The notebook is if you want to run it on google collab.